# specs/openapi/web/api.yaml
openapi: 3.0.3
info:
  title: Xilidan Web Gateway API
  description: REST API for the Xilidan web gateway service
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /organization:
    post:
      tags:
        - Organization
      summary: Create organization
      description: Create a new organization with positions and users
      operationId: createOrganization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Acme Corp
                positions:
                  type: array
                  items:
                    $ref: '#/components/schemas/PositionInput'
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    get:
      tags:
        - Organization
      summary: Get organization
      description: Retrieve organization details by user ID
      operationId: getOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: User ID to find their organization
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
        '500':
          description: Internal server error

    put:
      tags:
        - Organization
      summary: Update organization
      description: Update an existing organization
      operationId: updateOrganization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
              properties:
                id:
                  type: string
                  example: 550e8400-e29b-41d4-a716-446655440000
                name:
                  type: string
                  example: Acme Corp Updated
                positions:
                  type: array
                  items:
                    $ref: '#/components/schemas/PositionInput'
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        position_id:
          type: integer
          format: int64
          example: 1
        job:
          type: string
          example: Software Engineer
    UserInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
    Position:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Developer
        is_reviewer:
          type: boolean
          example: false
    PositionInput:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Developer
        is_reviewer:
          type: boolean
          example: false
    Organization:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Acme Corp
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'